+++ src/Game.cpp
@@ -12,12 +12,17 @@
 
 #include "../include/lights/SpotLight.h"
 #include "../include/lights/PointLight.h"
+
+////NEW LIBRAIRES TESTING
 #include "document.h"
 #include "random.hpp"
+#include "eventqueue.h"
 
 rapidjson::Document document;
 using Random = effolkronium::random_static;
 auto val = Random::get(-1, 1);
+eventpp::CallbackList<void (const std::string &, const bool)> callbackList;
+////NEW LIBRAIRES TESTING
 
 namespace Game {
     void ImGui();